app/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
openai==1.45.0
requests==2.32.2
python-dotenv==1.0.1
python-docx==1.1.2
pymongo==4.8.0
Pillow==10.4.0